+++ src/Holidays.php
@@ -7,20 +7,31 @@ use codemonauts\holidays\services\HolidaysService;
 use codemonauts\holidays\variables\HolidaysVariable;
 use Craft;
 use craft\base\Plugin;
+use craft\helpers\Cp;
 use craft\helpers\UrlHelper;
 use craft\web\twig\variables\CraftVariable;
 use Yasumi\Yasumi;
 use yii\base\Event;
 
 /**
- * Class Holidays
- *
  * @property HolidaysService holidays
- * @package codemonauts\holidays
  */
 class Holidays extends Plugin
 {
-    public $hasCpSettings = true;
+    /**
+     * @var \codemonauts\holidays\Holidays
+     */
+    public static Holidays $plugin;
+
+    /**
+     * @var \codemonauts\holidays\models\Settings|null
+     */
+    public static ?Settings $settings;
+
+    /**
+     * @inheritDoc
+     */
+    public bool $hasCpSettings = true;
 
     /**
      * @inheritDoc
@@ -29,11 +40,15 @@ class Holidays extends Plugin
     {
         parent::init();
 
+        self::$plugin = $this;
+
+        self::$settings = self::$plugin->getSettings();
+
         $this->setComponents([
             'holidays' => HolidaysService::class,
         ]);
 
-        Event::on(CraftVariable::class, CraftVariable::EVENT_INIT, function(Event $event) {
+        Event::on(CraftVariable::class, CraftVariable::EVENT_INIT, function (Event $event) {
             $variable = $event->sender;
             $variable->attachBehaviors([
                 HolidaysVariable::class,
@@ -44,7 +59,7 @@ class Holidays extends Plugin
     /**
      * @inheritDoc
      */
-    public function afterInstall()
+    public function afterInstall(): void
     {
         parent::afterInstall();
 
@@ -60,7 +75,7 @@ class Holidays extends Plugin
     /**
      * @inheritDoc
      */
-    protected function createSettingsModel()
+    protected function createSettingsModel(): Settings
     {
         return new Settings();
     }
@@ -68,16 +83,32 @@ class Holidays extends Plugin
     /**
      * @inheritDoc
      */
-    protected function settingsHtml()
+    protected function settingsHtml(): ?string
     {
         $providers = Yasumi::getProviders();
 
         ksort($providers);
 
-        return Craft::$app->getView()->renderTemplate('holidays/settings', [
-                'settings' => $this->getSettings(),
-                'providers' => $providers,
-            ]
-        );
+        return Cp::selectizeFieldHtml([
+            'label' => Craft::t('holidays', 'Country'),
+            'id' => 'defaultCode',
+            'name' => 'defaultCode',
+            'instructions' => Craft::t('holidays', 'Please, select the default country or country and subregion.'),
+            'value' => self::$settings->defaultCode,
+            'options' => $providers,
+            'required' => true,
+            'includeEnvVars' => true,
+        ]);
+    }
+
+    /**
+     * Returns the holidays component.
+     *
+     * @return HolidaysService
+     * @throws \yii\base\InvalidConfigException
+     */
+    public function getHolidays(): HolidaysService
+    {
+        return $this->get('holidays');
     }
 }