+++ src/models/Settings.php
@@ -9,12 +9,12 @@ class Settings extends Model
     /**
      * @var string Default country code in ISO 3166-2 (https://en.wikipedia.org/wiki/ISO_3166-2) notation.
      */
-    public $defaultCode = '';
+    public string $defaultCode = '';
 
     /**
      * @inheritdoc
      */
-    public function rules()
+    public function rules(): array
     {
         return [
             ['defaultCode', 'required'],