+++ CHANGELOG.md
@@ -1,5 +1,18 @@
-# Craft Yasumi Changelog
+# Craft Holidays Plugin Changelog
+
+## 2.0.0 - 2022-05-09
+
+### Added
+
+- Craft CMS 4 compatibility.
+
+### Changed
+
+- Requires Craft CMS >= 4.0
+- Update Yasumi package and provider to 2.5.0
 
 ## 1.0.0 - 2019-12-19
+
 ### Added
+- 
 - Initial release